refactor: util 패키지에 존재하던 LineStrategy 클래스를 domain 패키지로 이동
