+++ src/main/java/ladder/domain/Ladder.java
@@ -1,7 +1,5 @@
 package ladder.domain;
 
-import ladder.utils.LineStrategy;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
