+++ src/main/java/ladder/domain/LadderGame.java
@@ -1,7 +1,5 @@
 package ladder.domain;
 
-import ladder.utils.LineStrategy;
-
 import java.util.List;
 import java.util.stream.Collectors;
