+++ src/main/java/ladder/domain/Line.java
@@ -1,7 +1,5 @@
 package ladder.domain;
 
-import ladder.utils.LineStrategy;
-
 import java.util.Collections;
 import java.util.List;
