+++ src/test/java/ladder/utils/NonContinuousRandomLineStrategyTest.java
@@ -1,5 +1,6 @@
 package ladder.utils;
 
+import ladder.domain.NonContinuousRandomLineStrategy;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
